Add ESarsa and rework environments
Added the ESarsa algorithm, which can take in a ESarsa or QLearning spec
in its constructor.

Added agent specs, so that each agent constructor gets passed an
environment and an agent spec which uniquely defines each agent. This
way the Agent interface is consistent among all agents.

Moved environment specs to the same package and file as the agent specs.

GridWorld now implements the Environment interface.

Added documentation.