Add MountainCar and modify GetReward function
Added the discrete Mountain Car environment.

Modified the GetReward() function of a Task. Now, the GetReward()
function takes in the state, action, and next state. This was needed
because in some environments, the reward is determined by the next
state. Without this change, Task structs would have to re-compute the
next state in order to generate these kinds of rewards. Since the
environment already calculates the next state, it is easier and more
efficient to just pass this already calculated value to GetReward().