Transpose batch
Changed the functionality of the tile coder. When tile coding batches,
we now assume that each row is a feature and that columns are samples in
the batch. For example, if given a matrix of shape 3 x 10, this function
assumes that we have a batch of 10 samples, where each samples is 3
features.

This was done to improve performance. Since Go uses row-major order,
this change will result in the tile coder requiring to read a contiguous
region of memory to access a single feature among all samples in the
batch in order to tile-code. Before, each feature was a column. When
accessing a feature for each sample in the batch, this may have caused
Go to read multiple discontiguous regions of memory, which would be more
expensive.

When tile coding a batch, we access (as a vector) a single feature for
all samples in the batch, and calculate the tile-coded index based on
calculations of these features. Because Go uses row-major order and we
now set the rows of a matrix as the features, when selecting a feature
among all samples in the batch, Go will require a read to a contiguous
region of memory. If features were columns, Go would require reading
multiple discontiguous regions of memory in order to access this feature
among all samples.