Reset embedded env when creating TileCoding Env
Changed the NewTileCoding function so that it resets the embedded
environment when creating a new TileCoding struct. The documentation
says that an environment should be ready to go and return the first
timestep when created using the constructor. This change now reflects
this, because before, no timestep was returned when creating a new
TileCoding from its constructor, and the first timestep would not have
been tile coded (it would have been returned *only* when the embedded
environment was created, only the second - end timesteps would be tile
coded before).