Allow non-uniform tilings
Modified the code so that now, each tiling in a tile coder can have a
different number of tiles. Now, we could create a tile coder with 2
tilings, the first have size 2x2 tiles and the other have 16x4 tiles
e.g. This is reflected in the bins argument to a TileCoder. The length
of bins is the number of tilings, and the sub-slices are the tiles per
dimension. The bins parameter is a [][]int.