Add continuous mountain car
Added the continuous action version of Mountain Car. Changed up the
mountaincar package so that the base struct stores and tracks all
variables and methods which are shared between the discrete and
continuous action versions of Mountain Car. The Discrete and Continuous
structs now implement the individual functionality needed for only that
environment (e.g. Discrete updates using discrete actions, Continuous
with continuous actions) and embed a base environment.