Condense duplicated code
Moved all constants to the base Mountain Car file. Moved the calculation
of the next state to the base file as well. Now, the Discrete and
Continuous Mountain Car files just calculate the force from the selected
action, pass that force to the base env to calculate the next state
(this is done in the base file, which reduces code duplication), and
then updates the base environment with the new state.