Set weights key as constant
For the Weights() and SetWeights() methods of agents, now they keys are
set as constants in the policy package. For example.
linear/continuous/policy has the MeanWeightsKey which determines the key
in the map returned by Weights() whose value is the mean weights for any
policy that has mean weights.