Force learners to work with specific policies
Before, Learners could work with any policy. This is not a good idea,
because they rely on getting the weights of a specific type of policy
through a weights map. Now, the Learner's constructors take in a
specific kind of policy, for which the weights map is fixed and
expected. E.g. QLearning only takes in EGreedy policies, it cannot take
in any other kind of policy.