Begin EGreedyMLP implementation
Started an implementation of an EGreedyMLP policy. This implementation
is not yet completed and actually contains a learner struct as well
which takes care of learning (for testing purposes). The learner uses
target networks to learn the policy. If no target networks are needed,
then set the polyak averaging tau to 0.