Add EGreedy + DeepQ using Gorgonia neural nets
Added a new EGreedy policy which uses Gorgonia and neural network
function approximation. The EGreedy policy is a multi-headed MLP,
meaning that it takes in the state feature vector and predicts N action
values, where N = the number of actions allowed in the environment.

Also added the DeepQ learning algorithm. This algorithm is simply
Q-learning using neural networks for function approximation. It is *not*
DQN.