Add more efficient online sampler
For algorithms that use experience replay, a new struct onlineCache
efficiently implements an online experience replay buffer. In this case,
algorithms that use experience replay can easily be converted to an
online algorithm, and the onlineCache will ensure that all ER operations
are implemented efficiently. An online cache struct is not as efficient
as an onlineCache.