Add BatchSize to Selector and add error msg
Changed the Selectors to have a BatchSize() function. This makes it
easier for agents to get the batch size. Rather than first creating the
replay buffer, then getting the batch size (this may require the policy
to be constructed first which in turn may need the batch size), now the
agent can just get and set the batch size from the replay's sampler.

Changed the OnlineCache to report an error if the cache is empty and
Sample() is called.