Add experience replay errors
Added ExpReplayError which encapsulates errors unique to experience
replay. Now, agents can check to see if a reported error is because the
buffer has too few samples when sampling, and simply not update until
the buffer has enough samples. Agents can also check if an error is due
to the buffer having too few samples to sample (capacity < min
capacity). Before, this was not possible.