Create general neural nets
Created a new package that implements general neural networks that can
be used for any function approximator. This way, the same neural net
code can be used for critics, egreedy policies, and policy
parameterizations, and code will not need to be duplicated for each
policy or learner which needs its own function approximator.