Change Serializable interface to use gob
Changed the Serializable interface to implement gob's Encoder and
Decoder interfaces. This way, we can more easily use Go's builtins for
encoding and decoding use the gob package. Before, we required that an
object could be saved to any io.Writer, but now it will be the binary
representation of the struct.