Generate MLP from helper function
Created a new helper function that generates MLPs given an input node
and specs for hidden layers. This way, other networks can easily
construct MLPs (e.g. final layers of CNN or TreeMLP) and use the MLP
code instead of creating their own MLPs. In this way, all code that
needs to use MLPs can simply use the multiHeadMLP struct.

Added to the NeuralNet interface to reflect these changes. Because other
networks such as a TreeMLP can use the multiHeadMLP, we will need to be
able to clone the MLP from to another graph given another input. E.g. in
the TreeMLP, if we want to clone it, the new root network (which itself
is a multiHeadMLP) will be producing a different output, and the cloned
leaf networks will need to use this output, not the previous output or a
new output node.