Add Gaussian policy from tree mlp
Added a Gaussian policy from a tree mlp. Also had to change the NNPolicy
interface so that SelectAction() did not return action values. This is
important because policies shouldn't be able to tell the action values,
this should be for agents using the TDError() method. E.g. in
actor-critic, the actor has no idea about TD errors.