Store target/train nets as NeuralNets
Before for target and training networks that took in batches of data, we
used policies that allowed for batches of input, but then action
selection was not allowed. Now, these networks are NeuralNets instead of
NNPolicies, which makes more sense. A policy can't take a batch of
inputs in, but a neural network can.

Changed the NNPolicy interface by removing ClonePolicyWithBatch().
Policies should not ever take in batches. Instead, use CloneWithBatch()
which clones the NeuralNet.