Work on GaussianTreeMLP
Continued working on GaussianTreeMLP, which should be almost done now.
Added CloneWithBatch() back to NNPolicy, since we will need "Batch
policies" in order to learn policy gradient methods. E.g. SAC requires
to compute the log probability of actions selected by the policy for N
states in a batch. Before this was not possible, since we could only
have a batch network, not a batch policy and the policy computes the log
probabiltiy, not the network.