Debug Vanilla PG
Debugged vanilla PG a bit. I think the errors (may) have been ironed
out. I was performing gradient ascent on the ""critic"" loss function. So,
the TD errors for the advantages were calculated super incorrectly. Not
100% sure if all bugs are fixed though. Sometimes the agent performs
really well, and sometimes not. Might be just that I have to tune it
more.