Create Config Lists
Created configuration lists, which are structs specialized for storing
individual Config files. Also added a TypedConfigList, which stores the
agent type that the configuration list is for. Using the
TypedConfigList, we can JSON deserialize a configuration file for a
TypedConfigList. The deserialization process will use the type
information to deserialize the ConfigList into its correct concrete type
while keeping the type as ConfigList. E.g. when deserializing, it will
equivalent to something like:

var config ConfigList = NewCategoricalVanillaPGConfigList(...)

So that the type of the deserialized config list is ConfigList, but the
data has been deserialized into its correct concrete type. This is super
useful, because otherwise we would need to know beforehand what type we
were deserializing (e.g. CategoricalVanillaPGConfigList,
EGreedyDeepQConfigList, etc...) before deserialization occurred. Now, we
can just deserialize the JSON file, and it will take care of everything
automatically. We can simply deserialize and call
config.At(i).CreateAgent(...) to create the agent.