Add factory methods for creating exp replay
Added some factory methods for creating experience replayers and
selectors. This design choice was chosen because Config files must be
JSON serializable, and Selectors and ExperienceReplayers are not JSON
serializable. In order to make them serializable, their unexported
fields would have to be exported. Additionally, since these types are
both interface types, we would need some kind of container to ensure
that marshaling and unmarshaling their JSON representations could ensure
that they would be unmarshaled back into their concrete types (similar
to what we've done with TypedConfigList's, Solver's, and InitWFn's).

With the factory method way, Config files just need to store the
parameters of the experience replayer, which is much easier and results
in less code.