Debug DeepQ
Fixed a number of bugs in the DeepQ agent. Most notably:
    1. In the Observe() method, the input action was shadowed, causing
       the replay buffer to store all actions as action 0.
    2. Networks and policies shared the same computational graph.

I have not been able to get great performance of DeepQ on MountainCar or
Cartpole, but I think this is due to hyperparameter tuning. I was able
to get very good performance if I used the DeepQ implementation with
tile-coding to learn on MountainCar using Qlearning with linear function
approximation. This still used the DeepQ file, but we just did not add
any hidden layers, activations, etc. and used tile coding. Because it
works for Qlearning, I think that the file should be good for DeepQ as
well, but just due to other factors (e.g. DeepQ isn't great on
MountainCar anyway, hyperparameter tuning) I wasn't able to get good
performance. Also note, that this was over a few hours of testing
different hypers for ~20,000 timesteps manually. I did not do a
systematic tuning of hypers.