Add gridworld to env configs and add step limit
Added a default gridworld to the environment configs so that the default
Goal gridworld can be used in experiments. This gridworld is 5x5. Also
added a step limit ender to the gridworld.

Note that the envconfig file only lets this default gridworld be
created. If you want to change the gridworld up, you'll have to either
edit the envconfig package to change the size of the gridworld, or
you'll have to manually create a gridworld using the gridworld and
environment packages (very easy to do, just use the gridworld's
constructor and make a task for the gridworld).