Fix bug and increase efficiency
Fixed a bug where the next action stored in the ER buffer was equal to
the previous action. That is in the SARSA tuple, the second A == the
first A.

Changed up the implementation so that it is much faster. We no longer
are using sets, which caused the speed to slow dramatically with
increasing buffer size.