Add default Fifo remover exp replay and fix bug
Fixed a bug where when removing indices from the cache's inUseIndices,
only the last removed index was re-added to the empty indices slice.
This would cause the ER buffer to slowly shrink over time for some
Removers.

Added a new struct that implements the default experience replay
behaviour of using a FiFo remover of batch size 1. This is the default
behaviour, and we can take advantage of the implementation to increase
its efficiency dramatically.

The fifoRemove1Cache struct implements an ER buffer where, once the
buffer is full, a single element is removed in a Fifo manner. This
implementation is much more efficiency than the general cache
implementation, and it does not depend on the max capacity of the cache
for efficiency. That is, all operations on the fifoRemove1Cache are
O(1), while for the cache the operation to remove indices from the
inUseBuffer is O(n), since we first sample indices to remove, then
search through the cache's indices to remove them from the inUseIndices
slice and add them to the emptyIndices slice.