Add GaussianTreeMLP
Added a new struct GaussianTreeMLP which implements a Gaussian
distribution parameterized by a tree MLP (see the network package).
This struct can compute the log PDF of actions inputted to the policy,
but not the log pdf of actions taken by the policy immediately (so that
backprop can go "through" the actions). Because of this, this policy is
useful for VanillaPG or REINFORCE. A separate policy will be made for
other algorithms like SAC which require backproping "through" actions.