Finish acrobot implementation
Finished the acrobot implementation with its tasks. Added a new
FunctionEnder type which ends episodes when some function of a vector
returns true. In the case of acrobot, the function computes the height
and checks if the current environment state has the tip of the second
link above some set height.

Changed Ender types to return an Ender instead of the underlying
concrete type from the constructor.