Add getters and adjust Reset()
Added getter methods, which will be needed for some Tasks (since they
will end an episode or produce rewards based on the box2d lander or some
other part of the environment that is outside the scope of the state,
action, nextState observation vectors). For example, the nextState
vector does not encode if the legs have touched the ground, but the
AtGoal() of a task should return true if both legs have touched the
ground (note that the reward still depends only on the state, action,
and nextState, but the observation vector does not encode everything
about the state, yet the reward calculation needs to know everything
about the state).

Fixed the Reset() method. Before, it would return an empty timestep.
Now, it takes a single environmental step with a [0, 0] action and
returns the timestep generated. This is also what OpenAI does.