Begin implementation of Lunar Lander task.
Began implementing lunarLander tasks. Created a new lunarLanderTask
interface, which is a Task that gets a lunarLander registered with it.
This way, the task can call the lunar lander getters to generate the
AtGoal() function, generate rewards, end episodes appropriately etc. The
lunarLander struct will still work with normal Tasks, but if the Task is
a lunarLanderTask, then it will first register itself with the Task
before running any episodes.