Flip end cases
The lunar lander env has a bounded position of the lander for each
state. Because of this, one end case should never run, but we keep it
for generality. Therefore, I've moved this case to the end of the check
for all end cases so that if the environment did end, we won't check the
case at all. This will save compute a bit.