Add index tile coder
Added a new tile coding environment that returns indices of non-zero
components of the one-hot tile-coded state observation vector as state
features instead of the full one-hot vector. This can be used itself as
a feature vector, but it is also a better implementation of a tile
coder because by working with the indices themselves instead of the full
tile-coded vector (which may be billions of weights long), we can
achieve much greater computational efficiency.

Linear discrete agents were adapted to work with index tile coding
environments. These agents work with the indices themselves to update
specific weights, and do not treat the indices themselves as features.
In this way, they basically work as in the normal tile-coding case (with
full feature vectors), but instead of performing vector operations to
update, they work with the specific indices to increase computational
efficiency. In this way, the results of using an index tile coder or a
regular tile coder should be no different. The only difference is how
computationally efficient each implementation is. Each should achieve
the exact same results.