Remove DeepQ target policy
Removed DeepQ's target policy. This policy just added a lot of overhead
and reduced computational efficiency. It turns out that we never even
actually needed it, because we can simply turn the behaviour policy to
evaluation mode, which has the same effects as keeping a separate target
policy for action selection. In this way though, we don't need a target
policy which means that each time we step, we won't need to copy weights
from the training net to the target net.