Fix bug where Agent stored actions were clipped
Fixed a major bug that has been going on under the scenes for a long
while now. In the Online experiment, an Agent would select and return an
action to the experiment. The experiment would then pass that action to
the environment. Because actions are pointers (*mat.VecDense), this
could lead to major problems. Many agents store their actions in buffers
or fields after the Observe() method has been called, to which we send
the pointer action that the agent has chosen. The issue was that first
the environment would have to be stepped by calling Environment.Step().
Once this happened, many environments (in fact, all environments
implemented so far) clip the action to stay within some set range.
Because the action is a pointer, this resulted in the action passed back
to the agent with Observe() also being clipped. That is, the action that
the agent selected would be clipped and then saved for agent updating
later. This is **not** what we want! What we really want is for the
action to be clipped **only in the environment Step() function when
updating the enviornment** and we want the agent to keep its orginally
selected action (unclipped). This has now been fixed. In the Online
experiment, a copy of the action is sent to the environment now.

The reason that this is problematic is that many functions (such as
clipping) happen in place.