Speed up LAC when decay == 0
When decay == 0, this is equivalent to not using any traces. In such a
case, significant amounts of the code can be skipped over, and instead
of using traces and updating every single weight, we can simply update
the weights only at the indices of non-zero components of the tile coded
feature vector, instead of all the weights.

This significantly speeds up the runtime. E.g. lunar lander takes > 10s
to complete 5,000 timesteps without the speed up. With the speed up, it
takes ~36s to run 500,000 timesteps. Of course, eligibility traces
cannot be used to speed things up. On Mountain Car, every 100,000
timesteps, ~1 second is saved.