Fix bug resulting in segmentation fault
Fixed a bug that would result in a segmentation fault occassionally. The
issue was that we set the control in the mjData to be a pointer to some
region of memory. So, on the C side, MuJoCo had a pointer to some
memory, which was allocated with Go. Once the Go function ended, there
was no longer a pointer to that Go variable, and the Go garbage collector
cleaned it up, not knowing that the C side was still using the variable.

This has been fixed with a new C function which sets the control in the
mjData appropriately by **copying** all data in the Go slice to the C
array. Once Go cleans up the variable, C still retains a copy of the
values it needs. The function is called `setControl`.