Prepare gym package to use GoGym
Removed all the code in the gym package. The gym package is going to now
be implemented with [GoGym](https://github.com/samuelfneumann/GoGym),
which provides Go bindings for OpenAI's Gym using cgo and the Python C
API.

By using GoGym, we shoud see significant performance improvements
compared to using an HTTP server. Additionally, it is much easier to run
on cloud computing resources, since we don't have to worry about getting
access to a socket for the HTTP server (which may not even be possible
on some cloud resources). Additionally, the implementation will no
longer be restricted by the networking resources, and can easily be run
concurrently.