Remove `RewardSpec()` from Environment interface
Removed the `RewardSpec()`, `Min()`, and `Max()` methods from the
Environment interface since they did not add anything to the code, and
we never even needed them.

Kept these methods in the `classiccontrol` package, but the methods
are no longer part of the interface. Removed the methods from the
`box2d` and `mujoco` packages.