Stop panicing and return errors
Changed up the API so that the whole module panics less. The major
change is that the `Reset()` and `Step()` methods of environments no
longer panic but return `error`s instead, and so do the environment
constructors. All other changes (including adding more errors) were to
ensure that the rest of the code worked with the changes to these
methods.