Finish maze implementation
Finished the maze implementation. The environment now uses a one-hot
encoding of state, instead of (x, y) coordinates as the state
observations. In the future, we'll add a wrapper so that girdworlds and
mazes can return the (x, y) coords as observations instead.

The reason I've changed to using one-hot encodings, is that now the
environment can actually be easily treated as being in the tabular
setting, similarly to the gridworld package. It also adds some
consistency between the two packages. Now, both packages are treated as
tabular, and a single wrapper can be used to convert them to a function
approximation setting.