Increase experience replay efficiency
Added an option to disable tracking of the next action in the SARSA
tuple for experience replay. This is because many algorithms sample the
next action themselves, rather than taking it from the buffer. In such a
case, there is no need to track the next action, this will just reduce
efficiency.

Added concurrency for sampling from and adding to the buffer. The state,
action, next state, and next action are each sampled from or added to
the buffer in parallel.