Don't display progressbar until running experiment
Before, when an experiment was created the progress bar would be
displayed. Now, we wait until `Run()` is called on the experiment before
displaying the progress bar. This is useful because otherwise the
experiment would have to be created *after* anything else was displayed
on the screen. Now, an experiment can be made whenever, and it only
displays to the screen when instructed to do so.