Stop waiting when adding to buffer
Removed the wait when adding to an experience replay buffer. Since we
wait for concurrent processes to finish before every operation, there is
no need to wait for the experience replayer to finish adding to its
cache when `Add` is called. If another operation is called after `Add`
was called, that operation will wait until `Add` finishes running before
that operation runs. In other words, we were too concurrency safe
before. Now we are still concurrency safe, but the runtime has sped up.